refactor(frontend): extract wishlist list logic
